# [v6] plumbing: format/packfile, Optimise packfile delta processing