--- a/plumbing/format/objfile/reader.go
+++ b/plumbing/format/objfile/reader.go
@@ -25,6 +25,7 @@ type Reader struct {
 	zlib    io.Reader
 	zlibref sync.ZLibReader
 	hasher  plumbing.Hasher
+	closed  bool
 }
 
 // NewReader returns a new Reader reading from r.
@@ -113,6 +114,10 @@ func (r *Reader) Hash() plumbing.Hash {
 // Close releases any resources consumed by the Reader. Calling Close does not
 // close the wrapped io.Reader originally passed to NewReader.
 func (r *Reader) Close() error {
+	if r.closed {
+		return nil
+	}
+	r.closed = true
 	sync.PutZlibReader(r.zlibref)
 	return nil
 }