--- a/plumbing/format/packfile/parser.go
+++ b/plumbing/format/packfile/parser.go
@@ -70,10 +70,10 @@ func NewParser(data io.Reader, opts ...ParserOption) *Parser {
 
 		lm, ok := p.storage.(LowMemoryCapable)
 		p.lowMemoryMode = ok && lm.LowMemoryMode()
+	}
 
-		if p.scanner.seeker == nil {
-			p.lowMemoryMode = false
-		}
+	if p.scanner.seeker == nil {
+		p.lowMemoryMode = false
 	}
 	p.scanner.lowMemoryMode = p.lowMemoryMode
 	p.cache = newParserCache()