--- a/plumbing/format/packfile/parser_options.go
+++ b/plumbing/format/packfile/parser_options.go
@@ -26,13 +26,22 @@ func WithScannerObservers(ob ...Observer) ParserOption {
 	}
 }
 
-// WithHighMemoryUsage makes the parser optimise for speed rather than
-// for memory consumption. This is disabled by default.
+// WithHighMemoryMode optimises the parser for speed rather than
+// for memory consumption, making the Parser faster from an execution
+// time perspective, but yielding much more allocations, which in the
+// long run could make the application slower due to GC pressure.
+//
+// When the parser is being used without a storage, this is enabled
+// automatically, as it can't operate without it. Some storage types
+// may no support low memory mode (i.e. memory storage), for storage
+// types that do support it, this becomes an opt-in feature.
 //
 // When enabled the inflated content of all delta objects (ofs and ref)
 // will be loaded into cache, making it faster to navigate through them.
-func WithHighMemoryUsage() ParserOption {
+// If the reader provided to the parser does not implement io.Seeker,
+// full objects may also be loaded into memory.
+func WithHighMemoryMode() ParserOption {
 	return func(p *Parser) {
-		p.lowMemory = false
+		p.lowMemoryMode = false
 	}
 }