… blame_test.go
… plumbing/format/packfile/packfile_test.go
… plumbing/format/packfile/parser_test.go
… plumbing/format/packfile/scanner.go
… plumbing/format/packfile/types.go
… plumbing/object/blob.go
… plumbing/transport/http/dumb.go
… plumbing/transport/loader.go
… plumbing/transport/pack.go
… plumbing/transport/push.go
… storage/filesystem/config_test.go
… storage/filesystem/dotgit/dotgit_rewrite_packed_refs.go
… storage/filesystem/dotgit/dotgit_test.go
… storage/filesystem/dotgit/repository_filesystem_test.go
… storage/filesystem/object.go
… storage/filesystem/object_test.go
… storage/filesystem/storage.go
… submodule_test.go
… utils/ioutil/sync.go
… utils/sync/bytes.go
… utils/sync/bytes_test.go
… utils/sync/zlib.go
… utils/sync/zlib_test.go
… worktree.go
… worktree_status.go