The memory churn for processing delta objects was quite high, due to the performance optimisation which kept in memory all the inflated contents of each object.

The new default only keeps in memory the contents of the parents needed to resolve the contents of the delta being processed. The buffers used for all the contents are reused to avoid unnecessary churn.

Users can opt-out by using `WithHighMemoryUsage()` when creating the Parser. The memory churn drop from 50-90%, depending on the amount of delta objects in the repository and their sizes. Due to the decreased GC pressure, performance has improved between 10-40%, within the same conditions.

```
goos: linux
goarch: amd64
pkg: github.com/go-git/go-git/v6/plumbing/format/packfile
cpu: AMD Ryzen 7 PRO 8840HS w/ Radeon 780M Graphics
                                                                │ /tmp/before │             /tmp/after              │
                                                                │   sec/op    │   sec/op     vs base                │
Parse/https://github.com/git-fixtures/root-references.git-16      2.560m ± 1%   1.827m ± 1%  -28.62% (p=0.000 n=10)
Parse/https://github.com/git-fixtures/basic.git-16                1.955m ± 1%   1.581m ± 1%  -19.09% (p=0.000 n=10)
Parse/https://github.com/git-fixtures/basic.git#01-16             1.969m ± 1%   1.588m ± 0%  -19.32% (p=0.000 n=10)
Parse/https://github.com/git-fixtures/basic.git#02-16             1.927m ± 1%   1.545m ± 1%  -19.80% (p=0.000 n=10)
Parse/https://github.com/src-d/go-git.git-16                      323.1m ± 1%   322.9m ± 1%        ~ (p=0.739 n=10)
Parse/https://github.com/git-fixtures/tags.git-16                 81.76µ ± 1%   56.25µ ± 1%  -31.20% (p=0.000 n=10)
Parse/https://github.com/spinnaker/spinnaker.git-16               141.3m ± 1%   118.7m ± 0%  -16.01% (p=0.000 n=10)
Parse/https://github.com/jamesob/desk.git-16                      17.32m ± 1%   14.77m ± 1%  -14.71% (p=0.000 n=10)
Parse/https://github.com/cpcs499/Final_Pres_P.git-16              26.04µ ± 1%   18.31µ ± 1%  -29.67% (p=0.000 n=10)
Parse/https://github.com/github/gem-builder.git-16                2.011m ± 2%   1.492m ± 1%  -25.79% (p=0.000 n=10)
Parse/https://github.com/githubtraining/example-branches.git-16   377.8µ ± 1%   202.4µ ± 2%  -46.43% (p=0.000 n=10)
Parse/https://github.com/rumpkernel/rumprun-xen.git-16            132.3m ± 1%   116.8m ± 3%  -11.70% (p=0.000 n=10)
Parse/https://github.com/mcuadros/skeetr.git-16                   5.735m ± 2%   3.829m ± 1%  -33.23% (p=0.000 n=10)
Parse/https://github.com/dezfowler/LiteMock.git-16                4.710m ± 3%   4.027m ± 0%  -14.50% (p=0.000 n=10)
Parse/https://github.com/tyba/storable.git-16                     30.32m ± 3%   24.70m ± 1%  -18.53% (p=0.000 n=10)
Parse/https://github.com/toqueteos/ts3.git-16                     2.853m ± 3%   1.867m ± 1%  -34.57% (p=0.000 n=10)
Parse/#00-16                                                      474.2µ ± 2%   248.4µ ± 1%  -47.61% (p=0.000 n=10)
Parse/#01-16                                                      3.284m ± 1%   2.320m ± 1%  -29.34% (p=0.000 n=10)
Parse/#02-16                                                      492.6µ ± 1%   361.2µ ± 1%  -26.67% (p=0.000 n=10)
Parse/#03-16                                                      65.89µ ± 1%   43.02µ ± 3%  -34.72% (p=0.000 n=10)
geomean                                                           2.685m        1.988m       -25.96%

                                                                │  /tmp/before  │              /tmp/after              │
                                                                │     B/op      │     B/op      vs base                │
Parse/https://github.com/git-fixtures/root-references.git-16      1874.7Ki ± 0%   129.3Ki ± 0%  -93.10% (p=0.000 n=10)
Parse/https://github.com/git-fixtures/basic.git-16                841.46Ki ± 0%   96.17Ki ± 0%  -88.57% (p=0.000 n=10)
Parse/https://github.com/git-fixtures/basic.git#01-16             900.05Ki ± 0%   94.21Ki ± 0%  -89.53% (p=0.000 n=10)
Parse/https://github.com/git-fixtures/basic.git#02-16             802.52Ki ± 0%   87.48Ki ± 0%  -89.10% (p=0.000 n=10)
Parse/https://github.com/src-d/go-git.git-16                       56.01Mi ± 0%   22.40Mi ± 0%  -60.00% (p=0.000 n=10)
Parse/https://github.com/git-fixtures/tags.git-16                 263.32Ki ± 0%   67.53Ki ± 0%  -74.35% (p=0.000 n=10)
Parse/https://github.com/spinnaker/spinnaker.git-16                83.15Mi ± 0%   20.92Mi ± 0%  -74.85% (p=0.000 n=10)
Parse/https://github.com/jamesob/desk.git-16                       9.310Mi ± 0%   1.836Mi ± 0%  -80.28% (p=0.000 n=10)
Parse/https://github.com/cpcs499/Final_Pres_P.git-16              126.63Ki ± 0%   63.16Ki ± 0%  -50.12% (p=0.000 n=10)
Parse/https://github.com/github/gem-builder.git-16                1379.7Ki ± 0%   297.7Ki ± 0%  -78.42% (p=0.000 n=10)
Parse/https://github.com/githubtraining/example-branches.git-16   743.24Ki ± 0%   92.65Ki ± 0%  -87.53% (p=0.000 n=10)
Parse/https://github.com/rumpkernel/rumprun-xen.git-16             59.62Mi ± 0%   13.58Mi ± 0%  -77.21% (p=0.000 n=10)
Parse/https://github.com/mcuadros/skeetr.git-16                   6289.7Ki ± 0%   543.0Ki ± 0%  -91.37% (p=0.000 n=10)
Parse/https://github.com/dezfowler/LiteMock.git-16                1349.1Ki ± 0%   145.8Ki ± 0%  -89.19% (p=0.000 n=10)
Parse/https://github.com/tyba/storable.git-16                     17.884Mi ± 0%   3.915Mi ± 0%  -78.11% (p=0.000 n=10)
Parse/https://github.com/toqueteos/ts3.git-16                     2240.3Ki ± 0%   303.4Ki ± 0%  -86.46% (p=0.000 n=10)
Parse/#00-16                                                      1219.4Ki ± 0%   113.0Ki ± 0%  -90.73% (p=0.000 n=10)
Parse/#01-16                                                      3370.3Ki ± 0%   324.9Ki ± 0%  -90.36% (p=0.000 n=10)
Parse/#02-16                                                      301.83Ki ± 0%   91.82Ki ± 0%  -69.58% (p=0.000 n=10)
Parse/#03-16                                                      230.85Ki ± 0%   67.29Ki ± 0%  -70.85% (p=0.000 n=10)
geomean                                                            2.197Mi        377.1Ki       -83.24%

                                                                │ /tmp/before │             /tmp/after             │
                                                                │  allocs/op  │  allocs/op   vs base               │
Parse/https://github.com/git-fixtures/root-references.git-16      1.196k ± 0%   1.114k ± 0%  -6.86% (p=0.000 n=10)
Parse/https://github.com/git-fixtures/basic.git-16                 619.0 ± 0%    587.0 ± 0%  -5.17% (p=0.000 n=10)
Parse/https://github.com/git-fixtures/basic.git#01-16              582.0 ± 0%    564.0 ± 0%  -3.09% (p=0.000 n=10)
Parse/https://github.com/git-fixtures/basic.git#02-16              552.0 ± 0%    502.0 ± 0%  -9.06% (p=0.000 n=10)
Parse/https://github.com/src-d/go-git.git-16                      67.13k ± 0%   67.66k ± 0%  +0.79% (p=0.000 n=10)
Parse/https://github.com/git-fixtures/tags.git-16                  156.0 ± 0%    142.0 ± 0%  -8.97% (p=0.000 n=10)
Parse/https://github.com/spinnaker/spinnaker.git-16               106.8k ± 0%   104.9k ± 0%  -1.77% (p=0.000 n=10)
Parse/https://github.com/jamesob/desk.git-16                      12.22k ± 0%   12.23k ± 0%  +0.12% (p=0.000 n=10)
Parse/https://github.com/cpcs499/Final_Pres_P.git-16               68.00 ± 0%    66.00 ± 0%  -2.94% (p=0.000 n=10)
Parse/https://github.com/github/gem-builder.git-16                1.785k ± 0%   1.786k ± 0%  +0.06% (p=0.000 n=10)
Parse/https://github.com/githubtraining/example-branches.git-16    535.0 ± 0%    511.0 ± 0%  -4.49% (p=0.000 n=10)
Parse/https://github.com/rumpkernel/rumprun-xen.git-16            70.25k ± 0%   67.84k ± 0%  -3.42% (p=0.000 n=10)
Parse/https://github.com/mcuadros/skeetr.git-16                   5.335k ± 0%   4.937k ± 0%  -7.46% (p=0.000 n=10)
Parse/https://github.com/dezfowler/LiteMock.git-16                1.140k ± 0%   1.033k ± 0%  -9.39% (p=0.000 n=10)
Parse/https://github.com/tyba/storable.git-16                     25.37k ± 0%   25.16k ± 0%  -0.83% (p=0.000 n=10)
Parse/https://github.com/toqueteos/ts3.git-16                     2.443k ± 0%   2.311k ± 0%  -5.40% (p=0.000 n=10)
Parse/#00-16                                                       881.0 ± 0%    831.0 ± 0%  -5.68% (p=0.000 n=10)
Parse/#01-16                                                      2.862k ± 0%   3.065k ± 0%  +7.09% (p=0.000 n=10)
Parse/#02-16                                                       166.0 ± 0%    154.0 ± 0%  -7.23% (p=0.000 n=10)
Parse/#03-16                                                       142.0 ± 0%    129.0 ± 0%  -9.15% (p=0.000 n=10)
geomean                                                           1.888k        1.808k       -4.23%
```

Fixes https://github.com/go-git/go-git/issues/1188.

:warning: Depends on the PR #1502.