@runxiyu 2025-04-17
This may also fix https://github.com/go-git/go-git/issues/1451. I will test it against my workloads when I have time.